refactor: Remove graphql usage from frontend
