+++ packages/itmat-ui-react/proxy.conf.js
@@ -2,12 +2,6 @@ const API_SERVER = 'http://localhost:3333';
 
 module.exports = [
     {
-        context: ['/graphql'],
-        target: API_SERVER,
-        secure: false,
-        changeOrigin: true
-    },
-    {
         context: ['/trpc'],
         target: API_SERVER,
         secure: false,
